answer

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -10,7 +10,7 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned long int bits, count = 0;
+	unsigned int bits, count = 0;
 
 	bits = n ^ m;
